Sample DAF actions from hand-coded distributions

+++ temporal_policies/envs/pybullet/table/primitives.py
@@ -69,7 +69,7 @@ class Primitive(envs.Primitive, abc.ABC):
         if random.random() < 0.9:
             action = self.normalize_action(self.sample_action().vector)
             action = np.random.normal(loc=action, scale=0.05)
-            action = action.clip(self.action_space.low, self.action_space.high)
+            action = action.astype(np.float32).clip(self.action_space.low, self.action_space.high)
             return action
         else:
             return super().sample()
